Audio/HDA+AC97: Made the device emulation's internal [in|out]put FIFO buffer size configurable via extra data "VBoxInternal2/Audio/Device/BufSize[In|Out]Ms" (optional, defaults to 1000ms if not set). ticketoem2ref:36
svn:sync-xref-src-repo-rev: r142445

+++ src/VBox/Devices/Audio/DrvAudioCommon.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvAudioCommon.cpp 82968 2020-02-04 10:35:17Z knut.osmundsen@oracle.com $ */
+/* $Id: DrvAudioCommon.cpp 87436 2021-01-26 16:59:29Z andreas.loeffler@oracle.com $ */
 /** @file
  * Intermedia audio driver, common routines.
  *
@@ -1180,7 +1180,7 @@ uint32_t DrvAudioHlpCalcBitrate(const PPDMAUDIOPCMPROPS pProps)
  * @param   cbSize              Size (in bytes) to align.
  * @param   pProps              PCM properties to align size to.
  */
-uint32_t DrvAudioHlpBytesAlign(uint32_t cbSize, const PPDMAUDIOPCMPROPS pProps)
+uint32_t DrvAudioHlpBytesAlign(size_t cbSize, const PPDMAUDIOPCMPROPS pProps)
 {
     AssertPtrReturn(pProps, 0);
 
@@ -1197,7 +1197,7 @@ uint32_t DrvAudioHlpBytesAlign(uint32_t cbSize, const PPDMAUDIOPCMPROPS pProps)
  * @param   cbSize              Size (in bytes) to check alignment for.
  * @param   pProps              PCM properties to use for checking the alignment.
  */
-bool DrvAudioHlpBytesIsAligned(uint32_t cbSize, const PPDMAUDIOPCMPROPS pProps)
+bool DrvAudioHlpBytesIsAligned(size_t cbSize, const PPDMAUDIOPCMPROPS pProps)
 {
     AssertPtrReturn(pProps, 0);
